Bug 1704020 - Move searchStr into Record
